[dependencies] translating from vty and hscurses key codes to gtk; keypad works
More precisely, numeric keypad movement works on my computer with vty
backend on both xterm and gnome terminal and with hscurses on xterm.
Due to problems with gnome terminal and terminfo, diagonal walking fails.
We'd need hacking around 2-character key codes on our side to fix that
or we can wait until gnome terminal is fixed or start an xterm via script.