Better handling of keys and keybindings.
In particular, it's easier now to configure the keybindings. All
we need to do is to make the current keybindings part of the state
(ok, and perhaps abstract from the commands so that we can serialize
them, too).

Furthermore, the key help is now generated from the keybindings
structure that's used to dispatch on the input key to the command
handlers, thereby making it easier to keep the help in sync with
the commands that actually are available.