bugfix: recalculate perception more often
Monsters can open doors, and that affects the FOV of the player.
Previously, we've only recalculated FOV after the move of the
player. We now also recalculate it before the move of the player.
FOV of the player is also used to check whether monsters can see
the player, so the double calc seems like a good compromise for
now.