bugfix: no message on intermediate redraws.
The level is redrawn after monster moves, but the (combat)
messages should only be shown at the end. Otherwise, if too many
combat messages accumulate, we will have to confirm the same
message multiple times.

Note that an even better fix would be to print the messages
during intermediate redraws and partially clear it if the message
is longer than a line, after the confirmation.