ultra-hack: default config file gets included in the binary
In this way, we avoid the usual headaches with paths, OSes and permissions.
At startup it's always parsed, but a user config is used instead, if it exists.
TODO: throughout the code, lookup default values in the default config instead
of specifying them every time an option is queried. In this way, the various
game balance parameters are in sync and all in one place (the default config).
The user configs can be short and modify only some options, the rest is taken
from the default config. Diffs between user and default configs are easy.
The cost of all this is a bit of dynamic typing, and so runtime failures.
Can be limited by defining some types for sets of config parameters.