Tweak and reenable old debug commands
This is more buggy than regular mode, since server debug options sometimes
make the client suspect of cheating and so crash the game.
Also, with the current code structure it's not trivial to turn server
debug on, during an already started game, which would be more useful.
This is probably not worth the effort.