Revert "Check if the travis lock up is deterministic"
This reverts commit 213495867a7218fc28d5c7898522ecbd96bf4b5c.

No, the travis lock up is not deterministic, this time it didn't happen
with absolute the same execution, except that the initial RNG seed
was given as an argument, not from System.Random. So this is most
probably caused by a bug in libs or GHC 7.6.3, so I won't bother,
at least until I can reproduce it locally.