Start path fragments with as many as possible non-diagonal moves
Note: paths start at the actor's goal and end at the actor position.
The result is that the actor starts with diagonal moves that get him
as close to the next turn (or the goal) as possible in the Euclidean
metric, so that the path segments may end in non-diagonal moves,
that are less effective in reducing Euclidean distanse.
This is very good, because it eliminates loops when an actor
chases another and starts with an Euclidean-ineffective move
and then has to revert it, as in (1 chases 2):

                             O...............O..O......
                    .....O-----.O.OO1....----..........
                    ......-   -..........-..|..........
                    ......-----..O.....O.|..|O......O..
                    .............OO....2O----O.........

                             O.......1.......O..O......
                    .....O-----.O.OO.....----..........
                    ......-   -..........-..|..........
                    ......-----..O.....O2|..|O......O..
                    .............OO.....O----O.........