A failed attempt to separate the types of various kinds of items
The last straw was 'Kind.Id ItemKindAny', which would need to be
the kind of all items, but since the type argument to Kind.Id is a phantom,
I can't tell by inspecting the value which of the item kind roles
it plays so I can't map such a value to any of 'ItemKind x'
to look it up in game content. And I can't make Kind.Id
a type family or anything, since I put them inside vectors,
and deriving vector types is a pain.

Let's wait until the code evolves and try again much later.
For now, if there's enough item definitinos, I can define
Kind.Id a two bytes instead of one and this ought to be enough
for everybody.