Revert "Prefer Bfs paths that quickly become Euclidean-close to target"
This reverts commit d523cba9fa34a18de27c1c861f76d5687ae17762.

The goal is accomplished in a simpler, cheaper way by trying cardinal
directions first, as already done in preferredMoves. It's not as
accurate, but that just adds some fuzz and makes AI more interesting
and less predictable on the scale of the whole path (even if each
individual move is easier to predict).