Revert "Reduce save size by not computing benefit of trunks"
This reverts commit 5b6064d52db9b65219de40f85ff0adea5d440467.
The optimization was obviously wrong --- trunks of some thrown items
are first seen by some faction inside the projectile and those
that survive the impact then lack the benefit data.
I'd need to limit that to blast and big actor trunks,
but this is getting too complex for such a small gain.