Record nested macros on stack
All the macros are placed in a stack of Either macro buffers, since
macros can be nested. Bottom of stack is reserved for the user's in-game
macro buffer, so the stack is never empty.

We record keystrokes in the topmost Left macro buffer. At any time
there's at most one Right macro buffer, i.e. a buffer that's not
available to record to, but ready to be repeated from the macro bellow
it.