Fix nested macros
We handle each predefined macro locally. Whenever handling a macro, we
push onto the stack a new buffer containing all actions introduced by
the macro, and separate places for in-game macros recorded by the macro
and last action performed in the macro.

Since macros can be made of other predefined macros, we stack buffers on
top of each other. Buffers are being consumed from the top until theres
only the last one left, i.e. the user's buffer dedicated for handling
in-game macros. We pop a buffer when there's no actions pending to
handle in it.