Make the buggy finite trajectories infinite (if range infinite)
They were buggy, because when a projectile was displaced or teleported
the trajectory was unchanged and so could stick outside the level or,
conversely, could not reach a boundary of the level despite enough range
(coming from force of the throw, affected by weight, expressed in speed).
When they are infinite, I will need to take care to bound them
when drawing, etc., but the semantics after displacement or teleportation
will be correct and no special sanitizing code is needed when
handling displacement or teleportation.